+++ data/beaconrestapi/src/main/java/tech/pegasys/teku/beaconrestapi/handlers/v1/beacon/GetBlockHeader.java
@@ -42,7 +42,8 @@ import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.provider.JsonProvider;
 
 public class GetBlockHeader extends AbstractHandler implements Handler {
-  public static final String ROUTE = "/eth/v1/beacon/headers/{block_id}";
+  private static final String OAPI_ROUTE = "/eth/v1/beacon/headers/:block_id";
+  public static final String ROUTE = routeWithBracedParameters(OAPI_ROUTE);
   private final ChainDataProvider chainDataProvider;
 
   public GetBlockHeader(final DataProvider dataProvider, final JsonProvider jsonProvider) {
@@ -56,7 +57,7 @@ public class GetBlockHeader extends AbstractHandler implements Handler {
   }
 
   @OpenApi(
-      path = ROUTE,
+      path = OAPI_ROUTE,
       method = HttpMethod.GET,
       summary = "Get block header",
       tags = {TAG_BEACON},