+++ data/beaconrestapi/src/main/java/tech/pegasys/teku/beaconrestapi/handlers/v1/beacon/GetStateFork.java
@@ -43,7 +43,8 @@ import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.provider.JsonProvider;
 
 public class GetStateFork extends AbstractHandler implements Handler {
-  public static final String ROUTE = "/eth/v1/beacon/states/{state_id}/fork";
+  private static final String OAPI_ROUTE = "/eth/v1/beacon/states/:state_id/fork";
+  public static final String ROUTE = routeWithBracedParameters(OAPI_ROUTE);
   private final ChainDataProvider chainDataProvider;
 
   public GetStateFork(final DataProvider dataProvider, final JsonProvider jsonProvider) {
@@ -57,7 +58,7 @@ public class GetStateFork extends AbstractHandler implements Handler {
   }
 
   @OpenApi(
-      path = ROUTE,
+      path = OAPI_ROUTE,
       method = HttpMethod.GET,
       summary = "Get state fork",
       tags = {TAG_BEACON, TAG_VALIDATOR_REQUIRED},