+++ ethereum/statetransition/src/main/java/tech/pegasys/teku/statetransition/attestation/AttestationManager.java
@@ -239,10 +239,6 @@ public class AttestationManager extends Service
       return;
     }
 
-    if (attestation.isAggregate()) {
-      aggregateValidator.addSeenAggregate(attestation);
-    }
-
     notifyAttestationsToSendSubscribers(attestation);
     notifyAllValidAttestationsSubscribers(attestation);
     attestation.markGossiped();