+++ ethereum/statetransition/src/test/java/tech/pegasys/teku/statetransition/synccommittee/SyncCommitteeStateUtilsTest.java
@@ -27,7 +27,7 @@ import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.TestSpecFactory;
-import tech.pegasys.teku.spec.config.TestConfigLoader;
+import tech.pegasys.teku.spec.config.SpecConfigLoader;
 import tech.pegasys.teku.spec.datastructures.blocks.BeaconBlockHeader;
 import tech.pegasys.teku.spec.datastructures.blocks.SlotAndBlockRoot;
 import tech.pegasys.teku.spec.datastructures.blocks.StateAndBlockSummary;
@@ -43,7 +43,7 @@ class SyncCommitteeStateUtilsTest {
   private static final UInt64 ALTAIR_FORK_EPOCH = UInt64.ONE;
   private final Spec spec =
       TestSpecFactory.createAltair(
-          TestConfigLoader.loadConfig(
+          SpecConfigLoader.loadConfig(
               "minimal",
               builder ->
                   builder.altairBuilder(